refactor: logging service
