+++ Capgemini.Xrm.CdsDataMigrator/Capgemini.Xrm.CdsDataMigratorLibrary/Services/LogToFileService.cs
@@ -1,0 +1,23 @@
+﻿using System;
+using System.Threading;
+using System.Windows.Forms;
+using Capgemini.Xrm.CdsDataMigratorLibrary.Enums;
+
+namespace Capgemini.Xrm.CdsDataMigratorLibrary.Services
+{
+    public class LogToFileService : AbstractLoggerService
+    {
+        private readonly ILogManagerContainer logManagerContainer;
+
+        public LogToFileService(ILogManagerContainer logManagerContainer)
+        {
+            this.logManagerContainer = logManagerContainer;
+        }
+
+        protected override void WriteLine(string message, LogLevel logLevel)
+        {
+            var logMessage = $"{DateTime.Now:dd-MMM-yyyy HH:mm:ss} - {message}{Environment.NewLine}";
+            logManagerContainer.WriteLine(logMessage, logLevel);
+        }
+    }
+}
